Refactor: Split parser.rs into smaller modules (#16)
